Prevent server side rendering of schema viewer
